refactor(controller):  입력 처리 InputHandler로 분리 및 재입력 로직 통합
- 템플릿 메서드 패턴과 전략패턴 적용
- controller의 책임 줄이기
- InputHandler: inputPurchaseAmount(), inputWinningNumbers() 위임
