+++ src/main/java/lotto/controller/LottoController.java
@@ -1,7 +1,6 @@
 package lotto.controller;
 
 import java.util.List;
-import lotto.domain.BonusNumber;
 import lotto.domain.Lotto;
 import lotto.domain.LottoGame;
 import lotto.domain.LottoResult;
@@ -10,18 +9,18 @@ import lotto.domain.PurchaseAmount;
 import lotto.domain.RandomLottoNumberGenerator;
 import lotto.domain.WinningNumbers;
 import lotto.service.LottoShop;
-import lotto.util.InputParser;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class LottoController {
-    private final InputView inputView;
+    private final InputHandler inputHandler;
     private final OutputView outputView;
     private final LottoShop lottoShop;
 
     public LottoController() {
-        this.inputView = new InputView();
+        InputView inputView = new InputView();
         this.outputView = new OutputView();
+        this.inputHandler = new InputHandler(inputView, outputView);
         this.lottoShop = new LottoShop(new RandomLottoNumberGenerator());
     }
 
@@ -34,11 +33,16 @@ public class LottoController {
     }
 
     private LottoGame prepareGame() {
-        PurchaseAmount purchaseAmount = inputPurchaseAmount();
-        List<Lotto> lottos = purchaseLottos(purchaseAmount.getLottoQuantity());
-        WinningNumbers winningNumbers = inputWinningNumbers();
+        PurchaseAmount purchaseAmount = inputHandler.inputPurchaseAmount();
+        List<Lotto> lottos = lottoShop.buyLottos(purchaseAmount.getLottoQuantity());
+        WinningNumbers winningNumbers = inputHandler.inputWinningNumbers();
 
         return new LottoGame(purchaseAmount, lottos, winningNumbers);
+    }
+
+    private void printPurchaseInfo(LottoGame game) {
+        outputView.printPurchaseCount(game.getLottoQuantity());
+        outputView.printLottos(game.getPurchasedLottos());
     }
 
     private void printResult(LottoResult result) {
@@ -48,63 +52,4 @@ public class LottoController {
         ProfitRate profitRate = result.calculateProfitRate();
         outputView.printProfitRate(profitRate.getValue());
     }
-
-    private void printPurchaseInfo(LottoGame game) {
-        outputView.printPurchaseCount(game.getLottoQuantity());
-        outputView.printLottos(game.getPurchasedLottos());
-    }
-
-    private PurchaseAmount inputPurchaseAmount() {
-        while (true) {
-            try {
-                String input = inputView.readPurchaseAmount();
-                return PurchaseAmount.from(input);
-            } catch (IllegalArgumentException e) {
-                outputView.printErrorMessage(e.getMessage());
-            }
-        }
-    }
-
-    private List<Lotto> purchaseLottos(int quantity) {
-        List<Lotto> lottos = lottoShop.buyLottos(quantity);
-        outputView.printPurchaseCount(quantity);
-        outputView.printLottos(lottos);
-        return lottos;
-    }
-
-    private WinningNumbers inputWinningNumbers() {
-        Lotto winningLotto = inputWinningNumbersList();
-        BonusNumber bonusNumber = inputBonusNumber(winningLotto);
-        return new WinningNumbers(winningLotto, bonusNumber);
-    }
-
-
-    private Lotto inputWinningNumbersList() {
-        while (true) {
-            try {
-                List<String> numberStrings = inputView.readWinningNumbers();
-                List<Integer> numbers = numberStrings.stream()
-                        .map(InputParser::parseToInteger)
-                        .toList();
-
-                return new Lotto(numbers);
-            } catch (IllegalArgumentException e) {
-                outputView.printErrorMessage(e.getMessage());
-            }
-        }
-    }
-
-    private BonusNumber inputBonusNumber(Lotto winningNumbers) {
-        while (true) {
-            try {
-                String input = inputView.readBonusNumber();
-                int value = InputParser.parseToInteger(input);
-                return BonusNumber.of(value, winningNumbers);
-
-            } catch (IllegalArgumentException e) {
-                outputView.printErrorMessage(e.getMessage());
-            }
-        }
-    }
-
 }
